credential: add support for helper-specific state
TODO: docs, tests

It can be very helpful for credential helpers to have some sort of a
state between invocations, e.g. to preserve the fact that a username was
provided via the URL already, not interactively by the user.

To support this use case and others, we now remember unrecognized lines
quoted back by the helper, and write them again e.g. when calling the
helper to store the credential.

Co-authored-by: Matthew John Cheetham <mjcheetham@outlook.com>

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Matthew John Cheetham <mjcheetham@outlook.com>