wincred: ignore unknown lines (do not die)
It is the expectation that credential helpers be liberal in what they
accept and conservative in what they return, to allow for future growth
and evolution of the protocol/interaction.

All of the other helpers (store, cache, osxkeychain, libsecret,
gnome-keyring) currently ignore any credential lines that are not
recognised, whereas the Windows helper (wincred) instead dies.

Fix the discrepancy and ignore unknown lines in the wincred helper.