credential: support extra arbitrary properties
Alongside the "known" credential properties (protocol, host, path,
username, password, etc.), add support for arbitrary custom properties
that can be added by transport helpers or credential helpers.

This allows for transport-specific information to be passed to external
credential helpers to aid with fetching or generating new credentials.
For example, the HTTP transport can include information about response
headers, the HTTP response code or method.

Conversely, this also permits credential helpers to return extra
information to be echoed back in subsequent helper calls, or other
helpers.
For example, in response to a `get` call a credential helper can include
some state or context value that Git will then write back in the
subsequent `store` or `erase` commands.

Any property that is "known" will be parsed in to the existing fields in
the credential stucture; only unknown fields will fall in to this new
bucket.

We use a string map to store the key/values and disallow multiple
properties with the same name (case-insensitive).