http: read HTTP WWW-Authenticate response headers
Read and store the HTTP response headers made for a partiular request.

This will allow us to pass important authentication challenge
information to credential helpers or others that would otherwise have
been lost.

libcurl only provides us with the ability to read all headers recieved
for a particular request, including any intermediate redirect requests
or proxies. The lines returned by libcurl include HTTP status lines
delinating any intermediate requests such as "HTTP/1.1 200". We use
these lines to clear the strbuf of headers fields as we encounter them
in order to only capture the final response headers.

In the future [1] we may be able to leverage functions to read headers
from libcurl itself, but as of today we must do this ourselves.

The header fields are appended in the strbuf without any processing;
that will be up to the credential helpers.

[1] https://daniel.haxx.se/blog/2022/03/22/a-headers-api-for-libcurl/