http: read HTTP WWW-Authenticate response headers
Read and store the HTTP WWW-Authenticate response headers made for
a partiular request.

Duplicate headers fields are combined such that the values are
concatenated, in the same order, with a comma `,`. Header names are
normalised to lowercase.

According to RFC2616 Section 4.2 [1], concatenating headers in this
way is semantically equivalent to the duplicate header fields. The same
RFC section also confirms that header field names are case insensitive.

libcurl only provides us with the ability to read all headers recieved
for a particular request, including any intermediate redirect requests
or proxies. The lines returned by libcurl include HTTP status lines
delinating any intermediate requests such as "HTTP/1.1 200". We use
these lines to reset the strbuf of WWW-Authenticate header value as
we encounter them in order to only capture the final response headers.

[1] https://datatracker.ietf.org/doc/html/rfc2616#section-4.2