credential: add credtype field to struct credential
Introduce a new field to the credential structure that can be used by
credential helpers to tell Git what type of credential is being
returned.

An empty value should be treated as "basic" (username/password) for
backwards compatibility purposes.

The driver of this change is for credential helpers to be able to
specify other types of credential, such as Bearer tokens, for libcurl to
use.