http: set specific auth scheme depending on credential
Specify the correct authentication scheme or credential type when
authenticating the curl handle. If the new `authtype` field in the
credential structure is `NULL` or "Basic" then use the existing
username/password options. If the field is "Bearer" then use the OAuth
bearer token curl option. Otherwise, the `authtype` field is the
authentication scheme and the `password` field is the raw, unencoded
value.