credential: add "type" credential attribute
Introduce a new credential attribure "type" that is used to specify the
type of credential that is being returned.

If this attribute is omitted it is assumed to have the value "Basic",
which is to say that the credential attributes "username" and "password"
represent a basic username and password respectively.