http: consolidate request headers on request slot
There are a number of different instances where HTTP request headers
are constructed and set on the curl handle without preserving a local
copy.

Once a list of headers has been set on a curl handle it is no longer
possible to modify the list without a local copy, as curl takes a copy
of the list when calling `curl_easy_setopt(curl, CURLOPT_HTTPHEADER, x)`.
Futhermore, it is not possible to retrieve any set headers from the curl
handle after being set.

In future commits we will want to modify the HTTP request headers in
response to 401 responses, based on information from credential helpers.
After credential helpers have been consulted we wish to re-attempt the
request with all the same headers as before, but with some extra
authentication-specific fields.

In order to acheive this we standardise building the request header list
on the active request slot so that any common authentication handling
functions is free to modify the list before a request is made.
When a slot is requested from `get_request_slot(int)` any headers from
previous usages are free'd and the set of default headers is copied
ready for the caller to append to.

The request headers are set on the curl handle at the start of
`start_active_slot` (which all HTTP requests call), just prior to the
network request actually being made.