http: use correct auth scheme for HTTP requests
When a "type" attribute has been set on the credential for HTTP
authentication, use the correct corresponding "Authorization" header
scheme for the credential.

If the "type" attribute is not set, or if is set explicitly to "Basic"
then treat the username and password attributes as such and continue to
set the `CURLOPT_USERNAME` and `CURLOPT_PASSWORD` options respectively.

If "type" is "Bearer" then set the `CURLOPT_XOAUTH2_BEARER` option to
the value of the "password" attribute. We reuse the "password" attribute
for the bearer token value as this attribute is already considered
sensitive and we can leverage any obscurification/scrubbing logic
already implemented by credental helpers.
For bearer tokens, also set the `CURLOPT_HTTPAUTH` option to
`CURLAUTH_BEARER` to avoid an extra round-trip for curl to re-discover
the available authentication methods.

Finally, if "type" is set to anything else, construct an "Authorization"
request header with the scheme set to "type" and the value taken from
the password attribute. This enables credential helpers to leverage new
authentication schemes without needing to build explicit support in
either Git or curl.