fixup! release: create initial Windows installer build workflow
Probably in yet another patch to accommodate Meson (don't you love it
now?) the spaces in the `GIT-VERSION-FILE` are now gone. Where before,
its contents looked somewhat like this:

	GIT_VERSION = 2.48.1.vfs.0.0

they now look this this (note the removed spaces):

	GIT_VERSION=2.48.1.vfs.0.0

Well, so here is yet another other patch to accommodate Meson, so that
we can finally get a pre-release of v2.48.1.vfs.0.0 out.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>