fixup! release: build unsigned Ubuntu .deb package
We need to require a newer Ubuntu version because Git fails to build
with Ubuntu 16.04.

This is needed because Ubuntu 16.04 apparently has a horribly outdated
curl version and Git must at all cost drop support for it: 4c1b7e364e59
(Merge branch 'bc/drop-ancient-libcurl-and-perl', 2024-12-04). I just
wish that at times pragmatism would beat ideology, but it's not to be
had in this instance.

Unfortunately, it requires a bit more than just changing the Ubuntu
version, as the newer version would now ask interactively for the
timezone while setting up the `tzdata` package.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>